Compiles with -Wall and -Werror :)
Former-commit-id: 2568e0a42e28bb6e3672c4d59b51f88afc74044e [formerly 06a845d3e844322116b61825f6d3ce81f0f8f2a7]
Former-commit-id: f90ca4adb84fe61fe1b3584b23f439184cd8af52

+++ librapid/include/librapid/array/assignOps.hpp
@@ -30,7 +30,7 @@ namespace librapid::detail {
 
 		if (multiThread) {
 #pragma omp parallel for shared(lhs, function, vectorSize) default(none)                           \
-  num_threads(global::numThreads)
+  num_threads(static_cast<int>(global::numThreads))
 			for (int64_t index = 0; index < vectorSize; index += packetWidth) {
 				lhs.writePacket(index, function.packet(index));
 			}
